+++ lib/src/math/mathutils.dart
@@ -27,7 +27,7 @@ class StarMathUtils {
   ///Round to a double place where the second input corresponds to the amount
   ///of decimal places in the first double (eg 1 ==> x.x, 2 ==> x.xx, 3 ==> x.xxx).
   ///
-  ///decPlaces must be >= 0
+  ///```decPlaces``` must be >= 0
   double roundToDouble(double input, int decPlaces) {
     double parsedInp = 1.0;
     int fac = pow(10, decPlaces);
@@ -44,7 +44,7 @@ class StarMathUtils {
   ///Round up (ceil) to a double place where the second input corresponds to the amount
   ///of decimal places in the first double (eg 1 ==> x.x, 2 ==> x.xx, 3 ==> x.xxx).
   ///
-  ///decPlaces must be >= 0
+  ///```decPlaces``` must be >= 0
   double ceilToDouble(double input, int decPlaces) {
     double parsedInp = 1.0;
     int fac = pow(10, decPlaces);
@@ -61,7 +61,7 @@ class StarMathUtils {
   ///Round down (floor) to a double place where the second input corresponds to the amount
   ///of decimal places in the first double (eg 1 ==> x.x, 2 ==> x.xx, 3 ==> x.xxx).
   ///
-  ///decPlaces must be >= 0
+  ///```decPlaces``` must be >= 0
   double floorToDouble(double input, int decPlaces) {
     double parsedInp = 1.0;
     int fac = pow(10, decPlaces);
@@ -75,12 +75,12 @@ class StarMathUtils {
     }
   }
 
-  ///Returns whether given input is a power of two.
+  ///Returns whether given ```input``` is a power of two.
   bool isPowerOfTwo(num input) {
     return log(2, input).ceil() == log(2, input).floor();
   }
 
-  ///Returns log given base and input.
+  ///Returns log given ```base``` and ```input```.
   double log(num base, num input) {
     return math.log(input) / math.log(base);
   }
@@ -90,7 +90,7 @@ class StarMathUtils {
     return input % 1 == 0.0;
   }
 
-  ///Returns factorial given integer input.
+  ///Returns factorial given integer ```input```.
   int factorial(int input) {
     if (input < 0) {
       throw ArgumentError('factorial($input) is undefined for negative arguments.');
@@ -105,12 +105,12 @@ class StarMathUtils {
     return r;
   }
 
-  ///A list of provided factorials, where factorials[i] == i!
+  ///A list of provided factorials, where ```factorials[i] == i!```
   ///
-  ///This list will grow the more you use the provided factorial method with values > 11
+  ///This list will grow the more you use the provided ```factorial``` method with values > 11
   static List<int> factorials = [1, 1, 2, 6, 24, 120, 720, 5040, 40320, 362880, 3628800, 39916800, 479001600, 6227020800, 87178291200];
 
-  ///Returns true if x is within some tolerance of y
+  ///Returns true if ```x``` is within some tolerance of ```y```
   ///
   ///Ex. 1: fuzzyEquals(1, 4, 3) returns true
   ///
@@ -128,11 +128,11 @@ class StarMathUtils {
 
 
 
-  ///Tests if the number n is probably a prime.
+  ///Tests if the number ```n``` is probably a prime.
   ///
   ///This variant of the probabilistic prime test by Miller–Rabin is deterministic.
   ///
-  ///It has been verified to return correct results for all n < 341,550,071,728,321.
+  ///It has been verified to return correct results for all ```n``` < 341,550,071,728,321.
   bool isPrime(int n) {
     if (n == 2 || n == 3 || n == 5) {
       return true;
@@ -200,7 +200,7 @@ class StarMathUtils {
 
   /// Best method for computing the arithmetic mean.
   ///
-  /// The alternative (x + y) / 2 fails for large values, which this handles.
+  /// The alternative ```(x + y) / 2``` fails for large values, which this handles.
   int mean(int x, int y) {
     return (x & y) + ((x ^ y) >> 1);
   }